+++ src/main/kotlin/stickyWallet/events/PayEvent.kt
@@ -5,12 +5,13 @@ import org.bukkit.event.Event
 import org.bukkit.event.HandlerList
 import stickyWallet.accounts.Account
 import stickyWallet.currencies.Currency
+import java.math.BigDecimal
 
 class PayEvent(
     val currency: Currency,
     val payer: Account,
     val receiver: Account,
-    val amount: Double
+    val amount: BigDecimal
 ) : Event(), Cancellable {
 
     private val handlerList = HandlerList()
