+++ src/main/kotlin/stickyWallet/interfaces/DataHandler.kt
@@ -2,12 +2,13 @@ package stickyWallet.interfaces
 
 import stickyWallet.accounts.Account
 import stickyWallet.currencies.Currency
+import java.math.BigDecimal
 import java.util.UUID
 
 abstract class DataHandler(val name: String) : UsePlugin {
     abstract fun initialize(): Unit
 
-    abstract fun getTopList(currency: Currency, offset: Long, amount: Int): Map<String, Double>?
+    abstract fun getTopList(currency: Currency, offset: Long, amount: Int): Map<String, BigDecimal>?
 
     abstract fun loadCurrencies(): Unit
     abstract fun saveCurrency(currency: Currency): Unit
